Cirrus: Windows: install ppm deps with Yarn script
Effectively: install ppm's dependencies (and run its postinstall
scripts) with Yarn, not npm, on Windows.

This makes the installation of ppm's dependencies consistently use
Yarn, not npm, in Cirrus CI. Consistent for all OSes/arches we build
for in Cirrus.

(* Although, at some level the postinstall scripts are run via npm as
a subprocess of node as a subprocess of Yarn in a shell/console
sub-process... Yes, in my opinion, ppm's postinstall scripts are way
too complicated...)

ALSO: This fixes an oversight in PR 239 where the build script of the
wrong project (ppm instead of core) was being run at one point, again
on Windows.

So, this should make it so Pulsar's core dependencies are built for
the correct Electron version instead of built for Node... Oops!